# specs/002-coach-availability-impersonation/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PracticePerfect — Coach Invitations, Availability & Impersonation
  version: 1.3.0
  description: |
    Contract for feature 002-coach-availability-impersonation. It is version **1.3.0** of the same
    `/api/v1` contract feature 001 defined at 1.2.0, and it is **purely additive**: every operation
    below is new, and the only changes to existing response models are new optional fields
    (`CurrentUser.impersonation`, `CurrentUser.impersonation_ended`,
    `TrainerPlayerSummary.availability`, `TrainerPlayerSummary.availability_updated_at`). No
    operation is removed, and no existing field changes type or meaning, so the frontend compiled
    against 1.2.0 keeps working until it chooses to read the new fields (research.md R2-22).

    This file documents **only what 1.3.0 adds**. Operations defined in
    `specs/001-user-roles-admin/contracts/openapi.yaml` are unchanged and not repeated here, with
    the single exception of the two schema extensions above, which are given in full so the
    frontend types have one authoritative source.

    Three groups of operations:

    * **Coach invitations** (`coach-invitations` tag, plus `/trainer/coach-invitations` and
      `/trainer/coaches`) — a trainer invites a coach, tracks the invitation, and manages the
      roster; the invited person previews, registers, or accepts. US-01.08, FR-001 – FR-023.
    * **Availability** (`availability` tag) — a coach's own week, a player profile's week, and the
...

...

...
servers:
...
security:
  - sessionCookie: []
paths:
...
  /trainer/coach-invitations:
    get:
      tags: [trainer]
      summary: List the coach invitations this trainer has issued
      operationId: listCoachInvitations
      x-requirements: [FR-004, FR-009]
      description: |
        Trainer only, scoped to the caller's own invitations with no parameter that could widen it.
        Rows in the `superseded` state are never returned: a resend leaves one invitation per
        address, not two (FR-005). `state` is the *presented* state, which folds in expiry and the
        block annotation by the precedence in data-model.md §101.1 — a client never computes it.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: state
          in: query
          required: false
          description: Optional filter on the presented state.
          schema:
            type: string
            enum: [awaiting, accepted, expired, revoked, blocked]
      responses:
        '200':
          description: One page of invitations, newest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CoachInvitationPage' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      tags: [trainer]
      summary: Invite a coach by email address
      operationId: issueCoachInvitation
      x-requirements: [FR-001, FR-002, FR-003, FR-007, FR-008, FR-010, FR-023]
      description: |
        Issues a single-use invitation that stops working seven days after issue, and mails it to
        the address with the trainer's name and brand and the personal message when one was given.

        The response never reveals whether the address already holds an account (FR-008): the
        outcome is identical either way, and every conflict that depends on the invited person's
        account surfaces later, to that person, at acceptance.

        Refused with `409 coach_invitation_pending` when this trainer already has an invitation to
        the same address that is awaiting a response and not yet expired; the error body names the
        existing invitation so the client can offer resend or revoke (FR-007). Refused with
        `403 account_not_active` when the trainer's own account is not Active (FR-010).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CoachInvitationCreate' }
      responses:
        '201':
          description: The invitation, as the trainer's list will show it.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CoachInvitation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: An invitation to this address is already awaiting a response (FR-007).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CoachInvitationConflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /trainer/coach-invitations/{invitation_id}/resend:
    post:
      tags: [trainer]
      summary: Re-send an invitation, replacing it with a fresh one
      operationId: resendCoachInvitation
      x-requirements: [FR-005, FR-009, FR-010, FR-023]
      description: |
        Permitted while the invitation is awaiting a response or has expired. The current row is
        marked `superseded` and a new seven-day invitation is issued to the same address in the same
        transaction, so the previous link stops admitting anyone immediately and the trainer's list
        still shows exactly one invitation for that address (FR-005).

        The response is the **new** invitation. Refused with `422 invitation_not_resendable` for an
        invitation that is accepted or revoked.
      parameters:
        - $ref: '#/components/parameters/CoachInvitationId'
      responses:
        '201':
          description: The replacement invitation.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CoachInvitation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ActionNotPermitted' }
  /trainer/coach-invitations/{invitation_id}/revoke:
    post:
      tags: [trainer]
      summary: Revoke an invitation that is awaiting a response
      operationId: revokeCoachInvitation
      x-requirements: [FR-006, FR-009, FR-023]
      description: |
        After this, following the link is refused with a plain statement that it is no longer valid.
        Refused with `422 invitation_not_revocable` for an invitation that is already accepted,
        revoked, or expired — an expired invitation needs no revoking and saying so is clearer than
        pretending the action did something.
      parameters:
        - $ref: '#/components/parameters/CoachInvitationId'
      responses:
        '200':
          description: The invitation, now revoked.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CoachInvitation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ActionNotPermitted' }
  /trainer/coaches:
    get:
      tags: [trainer]
      summary: List the coaches on this trainer's roster
      operationId: listTrainerCoaches
      x-requirements: [FR-020, FR-034, FR-036]
      description: |
        Trainer only, scoped to the caller's own roster with no parameter that could widen it —
        the same shape as `GET /trainer/players`. Each row carries the coach's stated week, so the
        list can render the "Best times" summary without one request per row; the client formats the
        summary from the structured slots (research.md R2-12).
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: q
          in: query
          required: false
          description: Free-text filter on the coach's name or invited address.
          schema: { type: string, maxLength: 200 }
      responses:
        '200':
          description: One page of coaches.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrainerCoachPage' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /trainer/coaches/{coach_user_id}:
    delete:
      tags: [trainer]
      summary: End a coach's assignment to this trainer
      operationId: endCoachAssignment
      x-requirements: [FR-021, FR-022, FR-023]
      description: |
        The coach is on no roster afterwards, reaches none of this trainer's data, keeps their own
        account, profile, and stated times, and is free to accept another trainer's invitation.
        Idempotent in effect but not in status: a coach who is not on this trainer's roster is a
        404, because the caller is naming a resource they do not have.
      parameters:
        - $ref: '#/components/parameters/CoachUserId'
      responses:
        '204': { description: The assignment has ended. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ==========================================================================
  # Coach invitations — invited person's side (FR-011 – FR-019)
  # ==========================================================================
...
  /trainer/coaches/{coach_user_id}/availability:
    get:
      tags: [availability]
      summary: Read a coach's stated week, as their trainer
      operationId: getCoachAvailabilityAsTrainer
      x-requirements: [FR-034, FR-036, FR-037]
      description: |
        Read-only and scoped to the caller's own roster: a coach on another trainer's roster is a
        404 (FR-036). There is no matching write operation anywhere in the contract — stated times
        are the person's own (FR-037).
      parameters:
        - $ref: '#/components/parameters/CoachUserId'
      responses:
        '200':
          description: The coach's stated week, with the date it was last revised.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AvailabilityWeek' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /trainer/players/{profile_id}/availability:
    get:
      tags: [availability]
      summary: Read a player profile's stated week, as their trainer
      operationId: getPlayerAvailabilityAsTrainer
      x-requirements: [FR-034, FR-036, FR-037, FR-039]
      description: |
        Read-only and scoped to an **Active** association between the caller and this profile, which
        is what makes FR-039's "disclosure stops the moment an association ends" true with no
        separate step: the same query that authorizes the read is the one that would stop matching.
      parameters:
        - $ref: '#/components/parameters/PlayerProfileId'
      responses:
        '200':
          description: The profile's stated week, with the date it was last revised.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AvailabilityWeek' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ==========================================================================
  # Impersonation (FR-040 – FR-056)
  # ==========================================================================
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pp_session
      description: |
        Unchanged from feature 001. While an impersonation is live the same cookie resolves to the
        impersonated account; no second credential is issued (research.md R2-14).
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    CoachInvitationId:
      name: invitation_id
      in: path
      required: true
      description: An invitation issued by the calling trainer. Another trainer's is a 404.
      schema: { type: string, format: uuid }
...
    CoachUserId:
      name: coach_user_id
      in: path
      required: true
      description: |
        A coach on the calling trainer's own roster. A coach on another roster, or on none, is a 404
        rather than a 403 — the caller is naming a resource they do not have.
      schema: { type: string, format: uuid }
    PlayerProfileId:
      name: profile_id
      in: path
      required: true
      description: |
        A player profile. Under `/me/players` it must be on the caller's own account; under
        `/trainer/players` it must hold an Active association with the caller. Unreachable profiles
        are 404, as in feature 001.
      schema: { type: string, format: uuid }
  responses:
    Unauthenticated:
      description: No usable session cookie.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: The caller's role or ownership does not permit this action.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No such resource is reachable by this caller.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ActionNotPermitted:
      description: Understood, well-formed, and refused by a domain rule.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationError:
      description: One or more fields are invalid; each offending field is identified.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationErrorResponse' }
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: true
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
    ValidationErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [message, fields]
          properties:
            message: { type: string }
            fields:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [field, message]
                properties:
                  field: { type: string }
                  message: { type: string }

    # --- Coach invitations ---
    CoachInvitationCreate:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
          description: The invited address. Lower-cased on receipt.
        invitee_name:
          type: [string, 'null']
          minLength: 1
          maxLength: 200
          description: |
            Optional. `null` when not given — never `""`, which is a 422 (Principle VI).
        message:
          type: [string, 'null']
          minLength: 1
          maxLength: 2000
          description: Optional personal message, carried into the invitation email.
    CoachInvitationState:
      type: string
      enum: [awaiting, accepted, expired, revoked, blocked]
      description: |
        The **presented** state (data-model.md §101.1). `superseded` is not in this enum because a
        superseded invitation is never returned. `expired` and `blocked` are derived, not stored.
    CoachInvitation:
      type: object
      additionalProperties: false
      required:
        [id, invited_email, invitee_name, message, state, issued_at, expires_at, accepted_at,
         revoked_at, blocked_reason, coach]
      properties:
        id: { type: string, format: uuid }
        invited_email: { type: string, format: email }
        invitee_name: { type: [string, 'null'] }
        message: { type: [string, 'null'] }
        state: { $ref: '#/components/schemas/CoachInvitationState' }
        issued_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        accepted_at: { type: [string, 'null'], format: date-time }
        revoked_at: { type: [string, 'null'], format: date-time }
        blocked_reason:
          type: [string, 'null']
          enum: [role_not_coach, already_assigned, null]
          description: |
            Why an acceptance was refused (FR-019). `already_assigned` never carries or implies the
            identity of the trainer the coach already works with (FR-015).
        coach:
          oneOf:
            - $ref: '#/components/schemas/CoachSummary'
            - type: 'null'
          description: The coach who accepted, once one has. `null` in every other state.
    CoachInvitationPage:
      type: object
      additionalProperties: false
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/CoachInvitation' }
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
    CoachInvitationConflict:
      type: object
      additionalProperties: false
      required: [error]
      description: FR-007's 409 — the existing invitation is named so the client can act on it.
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, invitation]
          properties:
            code: { type: string, enum: [coach_invitation_pending] }
            message: { type: string }
            invitation: { $ref: '#/components/schemas/CoachInvitation' }
...
    CoachSummary:
      type: object
      additionalProperties: false
      required: [user_id, first_name, last_name, email, status, photo_url]
      properties:
        user_id: { type: string, format: uuid }
        first_name: { type: string }
        last_name: { type: string }
        email: { type: string, format: email }
        status: { type: string, enum: [active, inactive, deleted] }
        photo_url: { type: [string, 'null'] }
    TrainerCoachSummary:
      allOf:
        - $ref: '#/components/schemas/CoachSummary'
        - type: object
          additionalProperties: false
          required: [joined_at, availability, availability_updated_at]
          properties:
            joined_at: { type: string, format: date-time }
            availability:
              type: array
              description: |
                The coach's stated slots, so the list renders a summary without one request per row
                (research.md R2-12). Empty for "no times set".
              items: { $ref: '#/components/schemas/AvailabilitySlot' }
            availability_updated_at: { type: [string, 'null'], format: date-time }
    TrainerCoachPage:
      type: object
      additionalProperties: false
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/TrainerCoachSummary' }
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }

    # --- Availability ---
    AvailabilitySlot:
      type: object
      additionalProperties: false
      required: [day_of_week, start_minute, end_minute]
      description: |
        One range on one day. Minutes from midnight on a 15-minute grid (research.md R2-08):
        `17:00–20:00` is `{day_of_week: 0, start_minute: 1020, end_minute: 1200}`.
      properties:
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
          description: 0 = Monday … 6 = Sunday.
        start_minute:
          type: integer
          minimum: 0
          maximum: 1425
          multipleOf: 15
        end_minute:
          type: integer
          minimum: 15
          maximum: 1440
          multipleOf: 15
          description: May be 1440 (midnight). Must be greater than `start_minute`.
    AvailabilityWeek:
      type: object
      additionalProperties: false
      required: [slots, updated_at]
      properties:
        slots:
          type: array
          maxItems: 42
          description: |
            Ordered by `(day_of_week, start_minute)`. At most six per day (FR-028), so at most 42.
          items: { $ref: '#/components/schemas/AvailabilitySlot' }
        updated_at:
          type: [string, 'null']
          format: date-time
          description: |
            `null` means never stated. A non-null value with an empty `slots` means deliberately
            cleared. Neither means "unavailable" (FR-035).
...